Reorganize public members in filesystem package, other type hint fixes
