+++ modguard/filesystem/__init__.py
@@ -1,16 +1,3 @@
-from modguard import Boundary, public
-from .service import (
-    get_cwd,
-    chdir,
-    canonical,
-    read_file,
-    write_file,
-    parse_ast,
-    walk_pyfiles,
-    walk_pypackages,
-    file_to_module_path,
-    module_to_file_path,
-)
+from modguard import Boundary
 
 Boundary()
-public()
