+++ modguard/filesystem/interface.py
@@ -1,0 +1,18 @@
+from modguard import public
+
+# ruff: noqa: F401
+# pyright: reportUnusedImport=false
+from .service import (
+    get_cwd,
+    chdir,
+    canonical,
+    read_file,
+    write_file,
+    parse_ast,
+    walk_pyfiles,
+    walk_pypackages,
+    file_to_module_path,
+    module_to_file_path,
+)
+
+public()
