+++ modguard/init.py
@@ -3,10 +3,11 @@ from enum import Enum
 import os
 from typing import Optional
 
-from modguard import errors, filesystem as fs
+from modguard import errors
+from modguard.filesystem import interface as fs
 from modguard.check import check_import
 from modguard.core import PublicMember
-from modguard.parsing.boundary import add_boundary, has_boundary, build_boundary_trie
+from modguard.parsing.boundary import add_boundary, build_boundary_trie
 from modguard.parsing.imports import get_imports
 from modguard.parsing.public import mark_as_public
